requirements_v3_1.txt
#############################################################
# REQUIREMENTS v3.1 - All Dependencies
#############################################################

# Audio Processing
librosa==0.10.0
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.1

# ML & Models
torch==2.0.1
faster-whisper==0.10.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
onnxruntime==1.16.0

# Configuration
PyYAML==6.0
pydantic==2.0.0

# Web Framework & Async
fastapi==0.103.0
uvicorn==0.23.2
aiofiles==23.1.0
aioredis==2.0.1  # FIX #9: Redis queue

# Metrics & Monitoring
prometheus-client==0.18.0

# Database (optional)
sqlalchemy==2.0.0
psycopg2-binary==2.9.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Logging & Observability
python-json-logger==2.0.7

# Code Quality
black==23.9.1
flake8==6.0.0
mypy==1.4.1
isort==5.12.0

# Security
requests==2.31.0
cryptography==41.0.0

# Development
ipython==8.15.0
jupyter==1.0.0

# Production Server (optional)
gunicorn==21.2.0

# Cloud Integrations (optional)
boto3==1.28.0  # AWS CloudWatch
google-cloud-logging==3.5.0  # GCP Logging
